Eigenvalue Problem solver (EPS)

SlepcWrap.create_eps — Method

create_eps(A::PetscMat, B::PetscMat)

For a generalized eigenvalue problem

SlepcWrap.create_eps — Method

create_eps(A::PetscMat)

For a standard eigenvalue prolem.

SlepcWrap.eigenvalues2file — Function

eigenvalues2file(eps::SlepcEPS, ieigs, eigs_path = "eigenvalues.dat"; mpi_rank = 0, delim=",")

Write eigenvalues to a CSV file.

If two_cols == true, the output file is a (neig, 2) array (without any complex number).

mpi_rank is the rank of the processor writing the eigenvalue file.

SlepcWrap.eigenvectors2file — Function

eigenvectors2file(eps::SlepcEPS, ivecs, vectors_path = "eigenvectors"; mpi_rank = 0, type = "ascii", format = PETSC_VIEWER_ASCII_CSV)

Concatenate specified eigenvectors in two files : real and imag parts.

Warning

This function is experimental : it may allocate a lot of memory. Use it at your own risk.

SlepcWrap.eigenvectors2mat — Function

eigenvectors2mat(eps::SlepcEPS, ivecs)

Concatenate specified eigenvectors in two matrices : real and imag parts.

SlepcWrap.get_eig — Method

get_eig(eps::SlepcEPS, ieigs)

Get the ieig-th eigenvalues as a complex array. Starts at 1 (Julia 1-based indexing)

SlepcWrap.get_eig — Method

get_eig(eps::SlepcEPS, ieig::Integer)

Get the ieig-th eigenvalue as a complex number. Starts at 1 (Julia 1-based indexing)

SlepcWrap.get_eigenpair — Method

get_eigenpair(eps::SlepcEPS, ieig)

Return the ieig-th eigenpair (eigenvalue, eigenvector). ieig starts at 1 (Julia 1-based indexing)

SlepcWrap.get_eigenvector — Method

get_eigenvector(eps::SlepcEPS, ivec)

Return a tuple (vecr, veci) corresponding to the ivec-th eigenvector. ivec starts at 1 (Julia 1-based indexing)

SlepcWrap.get_eigs — Method

get_eigs(eps::SlepcEPS)

Returns all the converged eigenvalues as one complex array