Initialization

SlepcWrap.SlepcFinalize — Method

Wrapper to SlepcFinalize

SlepcWrap.SlepcInitialize — Function

Initialize SLEPc.

If cmd_line_args == true, then command line arguments passed to Julia are used as arguments for SLEPc (leading to a call to SlepcInitializeNoPointers).

Otherwise, if cmd_line_args == false, initialize SLEPc without arguments (leading to a call to SlepcInitializeNoArguments).

SlepcWrap.SlepcInitialize — Method

Wrapper to SlepcInitializeNoPointers

Implementation

I don't know if I am supposed to use PetscInt or not...